Create orderly remote

Implements an orderly "remote" using OrderlyWeb as a backend. Use this within an orderly_config.yml configuration.

orderlyweb_remote(host, port, token, https = TRUE, prefix = NULL,
  name = NULL)

Arguments

host

Fully qualified hostname for the OrderlyWeb instance

port

Port to use

token

Your application token for authentication. The appropriate value here will depend on the authentication support that is built into the OrderlyWeb server that you are communicating with. Provide the token directly (as a string) or provide a callback function that takes no arguments and returns token.

https

Optional logical, indicating if this is an https connection - this should be TRUE in all production settings or credentials will be sent in the clear!

prefix

A prefix, if your OrderlyWeb server is mounted at a path within some larger website.

name

A friendly name for the server (e.g, "production" or "testing") which may be printed when using the remote, or when authenticating. If not provided then a name will be constructed from host, port and (if provided) prefix.

Examples

remote <- orderlyweb::orderlyweb_remote("example.com", 443, "mytoken")
remote
#> <orderlyweb_remote>
#>   Public:
#>     bundle_import: function (path, progress = TRUE) 
#>     bundle_pack: function (name, parameters = NULL, instance = NULL, progress = TRUE) 
#>     initialize: function (host, port, token, https, prefix, name) 
#>     kill: function (key) 
#>     list_reports: function () 
#>     list_versions: function (name) 
#>     metadata: function (name, id) 
#>     name: example.com:443
#>     pull: function (name, id, progress = TRUE) 
#>     queue_status: function () 
#>     run: function (name, parameters = NULL, ref = NULL, timeout = NULL, 
#>     url_report: function (name, id) 
#>   Private:
#>     client: orderlyweb, R6